+++ scripts/config/constants.config.mjs
@@ -35,6 +35,7 @@ export const SETTING_KEYS = {
   HOVER_LABEL_Y_OFFSET: "hoverLabelYOffset",
   HOVER_LABEL_X_OFFSET: "hoverLabelXOffset",
   GRIDLESS_DISTANCE_MODE: "gridlessDistanceMode",
+  GRIDLESS_TOKEN_SHAPE: "gridlessTokenShape",
   CREATURES_PRONE: "proneCreatures"
 };